indexer/requirements.txt
requests
PyYAML
together
angle-emb
pandas
python-dotenv
